refactor: Move socialIcons folder into features/auth
