+++ src/shared/ui/index.ts
@@ -1,4 +1,3 @@
-import { SocialIcons } from './socialIcons/SocialIcons'
 // Imports
 export { Button } from './button'
 export { Card } from './card'
@@ -15,7 +14,7 @@ export { Recaptcha } from './reCaptcha/index'
 export { Scroll } from './scroll'
 export { SelectBox, SelectItem } from './selectBox'
 export { Sidebar } from './sidebar'
-export { SocialIcons } from './socialIcons'
+export { SocialIcons } from '../../features/auth/ui/socialIcons'
 export { Tabs } from './tabs'
 export { TextField } from './textField'
 export { Typography } from './typography'
